feat: improve theme
Use the complementary color to the one commonly used for the blindnet
and privateform brands (#dc4c87 on blindnet.io) with the default dark
theme to make the content more readable and impactful, and to make the
two brands work better together if needed.

For the light theme, this color has been shifted to a darker tone using
a palette generator and try-and-error on https://docusaurus.io/docs/styling-layout
in order to ensure a good enough contrast for readability.

Finally, some wordings has also been updated for consistency with
previous changes and better impact.